mobilitydb/requirements.txt
python-decouple
pathlib 
requests
zipfile36
pandas
psycopg2